# packages/schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Jordanian Digital Heritage Platform API",
    "description": "The core layer of the Jordanian Digital Heritage Platform. Protected content is served as watermarked tiles only; OCR text is internal and never returned.",
    "version": "0.1.0"
  },
  "paths": {
    "/works": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "List Works",
        "operationId": "list_works_works_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "access_class",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AccessClass"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Class"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paginated_WorkSummary_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "catalog"
        ],
        "summary": "Create Work",
        "operationId": "create_work_works_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/works/{name}": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Get Work",
        "operationId": "get_work_works__name__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkDetail"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "catalog"
        ],
        "summary": "Update Work",
        "operationId": "update_work_works__name__patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/collections": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "List Collections",
        "operationId": "list_collections_collections_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paginated_CollectionSummary_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/vocabulary/{scheme}": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "List Terms",
        "operationId": "list_terms_vocabulary__scheme__get",
        "parameters": [
          {
            "name": "scheme",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/TermScheme"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paginated_TermSummary_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccessClass": {
        "type": "string",
        "enum": [
          "open",
          "registered",
          "paid",
          "restricted",
          "embargoed"
        ],
        "title": "AccessClass"
      },
...
      "Citation": {
        "properties": {
          "ark": {
            "type": "string",
            "title": "Ark"
          },
          "ar": {
            "type": "string",
            "title": "Ar"
          },
          "en": {
            "type": "string",
            "title": "En"
          }
        },
        "type": "object",
        "required": [
          "ark",
          "ar",
          "en"
        ],
        "title": "Citation"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Paginated_CollectionSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CollectionSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Paginated[CollectionSummary]"
      },
...
      "Paginated_TermSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TermSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Paginated[TermSummary]"
      },
      "Paginated_WorkSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Paginated[WorkSummary]"
      },
...
      "PublishState": {
        "type": "string",
        "enum": [
          "draft",
          "review",
          "published",
          "withdrawn"
        ],
        "title": "PublishState"
      },
...
      "TermScheme": {
        "type": "string",
        "enum": [
          "lcsh",
          "local_ar",
          "gazetteer_jo",
          "period_jo",
          "material",
          "theme"
        ],
        "title": "TermScheme"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkCreate": {
        "properties": {
          "title_ar": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Title Ar"
          },
          "title_translit": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Title Translit"
          },
          "title_en": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Title En"
          },
          "uniform_title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Uniform Title"
          },
          "language": {
            "type": "string",
            "maxLength": 3,
            "minLength": 3,
            "title": "Language",
            "default": "ara"
          },
          "script": {
            "type": "string",
            "maxLength": 4,
            "minLength": 4,
            "title": "Script",
            "default": "Arab"
          },
          "date_edtf": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Edtf"
          },
          "date_hijri": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Hijri"
          },
          "description_ar": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description Ar"
          },
          "description_en": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description En"
          },
          "extent": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Extent"
          },
          "rights_statement": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Rights Statement"
          },
          "rights_basis": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Rights Basis"
          },
          "access_class": {
            "$ref": "#/components/schemas/AccessClass",
            "default": "registered"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title_ar"
        ],
        "title": "WorkCreate"
      },
      "WorkDetail": {
        "properties": {
          "public_id": {
            "type": "string",
            "title": "Public Id"
          },
          "ark": {
            "type": "string",
            "title": "Ark"
          },
          "title_ar": {
            "type": "string",
            "title": "Title Ar"
          },
          "title_translit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title Translit"
          },
          "title_en": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title En"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "script": {
            "type": "string",
            "title": "Script"
          },
          "date_edtf": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Edtf"
          },
          "date_hijri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Hijri"
          },
          "access_class": {
            "$ref": "#/components/schemas/AccessClass"
          },
          "publish_state": {
            "$ref": "#/components/schemas/PublishState"
          },
          "page_count": {
            "type": "integer",
            "title": "Page Count"
          },
          "sample_page_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Page Limit"
          },
          "thumbnail_available": {
            "type": "boolean",
            "title": "Thumbnail Available"
          },
          "uniform_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uniform Title"
          },
          "description_ar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description Ar"
          },
          "description_en": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description En"
          },
          "extent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extent"
          },
          "rights_statement": {
            "type": "string",
            "title": "Rights Statement"
          },
          "rights_basis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rights Basis"
          },
          "pricing": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pricing"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "agents": {
            "items": {
              "$ref": "#/components/schemas/AgentRef"
            },
            "type": "array",
            "title": "Agents"
          },
          "terms": {
            "items": {
              "$ref": "#/components/schemas/TermRef"
            },
            "type": "array",
            "title": "Terms"
          },
          "collections": {
            "items": {
              "$ref": "#/components/schemas/CollectionRef"
            },
            "type": "array",
            "title": "Collections"
          },
          "provenance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provenance"
          },
          "citation": {
            "$ref": "#/components/schemas/Citation"
          },
          "can_read": {
            "type": "boolean",
            "title": "Can Read"
          },
          "can_request_access": {
            "type": "boolean",
            "title": "Can Request Access"
          }
        },
        "type": "object",
        "required": [
          "public_id",
          "ark",
          "title_ar",
          "title_translit",
          "title_en",
          "language",
          "script",
          "date_edtf",
          "date_hijri",
          "access_class",
          "publish_state",
          "page_count",
          "sample_page_limit",
          "thumbnail_available",
          "uniform_title",
          "description_ar",
          "description_en",
          "extent",
          "rights_statement",
          "published_at",
          "agents",
          "terms",
          "collections",
          "provenance",
          "citation",
          "can_read",
          "can_request_access"
        ],
        "title": "WorkDetail"
      },
      "WorkSummary": {
        "properties": {
          "public_id": {
            "type": "string",
            "title": "Public Id"
          },
          "ark": {
            "type": "string",
            "title": "Ark"
          },
          "title_ar": {
            "type": "string",
            "title": "Title Ar"
          },
          "title_translit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title Translit"
          },
          "title_en": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title En"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "script": {
            "type": "string",
            "title": "Script"
          },
          "date_edtf": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Edtf"
          },
          "date_hijri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Hijri"
          },
          "access_class": {
            "$ref": "#/components/schemas/AccessClass"
          },
          "publish_state": {
            "$ref": "#/components/schemas/PublishState"
          },
          "page_count": {
            "type": "integer",
            "title": "Page Count"
          },
          "sample_page_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Page Limit"
          },
          "thumbnail_available": {
            "type": "boolean",
            "title": "Thumbnail Available"
          }
        },
        "type": "object",
        "required": [
          "public_id",
          "ark",
          "title_ar",
          "title_translit",
          "title_en",
          "language",
          "script",
          "date_edtf",
          "date_hijri",
          "access_class",
          "publish_state",
          "page_count",
          "sample_page_limit",
          "thumbnail_available"
        ],
        "title": "WorkSummary"
      },
      "WorkUpdate": {
        "properties": {
          "title_ar": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title Ar"
          },
          "title_translit": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Title Translit"
          },
          "title_en": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Title En"
          },
          "uniform_title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Uniform Title"
          },
          "language": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 3,
                "minLength": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "script": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4,
                "minLength": 4
              },
              {
                "type": "null"
              }
            ],
            "title": "Script"
          },
          "date_edtf": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Edtf"
          },
          "date_hijri": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Hijri"
          },
          "description_ar": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description Ar"
          },
          "description_en": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description En"
          },
          "extent": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Extent"
          },
          "rights_statement": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Rights Statement"
          },
          "rights_basis": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Rights Basis"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "WorkUpdate",
        "description": "Partial update. Access class and pricing are not here: they need a second approver."
      }
    }
  }
}